+++ src/main/java/com/example/noticeboardapi/domain/comment/repository/TreePathCommandRepository.java
@@ -16,8 +16,8 @@ public class TreePathCommandRepository {
 
     public void saveCommentTreePath(Long postNo, Long commentNo) {
         dslContext.insertInto(TREE_PATH, TREE_PATH.POST_ID, TREE_PATH.ANCESTOR, TREE_PATH.DESCENDANT, TREE_PATH.DEPTH)
-                .values(postNo, commentNo, commentNo, 0)
-                .values(postNo, ROOT_COMMENT_ID, commentNo, 1)
+                .values(postNo, ULong.valueOf(commentNo), ULong.valueOf(commentNo), 0)
+                .values(postNo, ULong.valueOf(ROOT_COMMENT_ID), ULong.valueOf(commentNo), 1)
                 .execute();
     }
 
@@ -26,8 +26,8 @@ public class TreePathCommandRepository {
                 .select(DSL.select(DSL.val(postNo), TREE_PATH.ANCESTOR,
                                 DSL.val(commentNo), TREE_PATH.DEPTH.plus(1))
                         .from(TREE_PATH)
-                        .where(TREE_PATH.DESCENDANT.eq(parentCommentNo))
-                        .unionAll(DSL.select(DSL.val(postNo), DSL.val(commentNo),
+                        .where(TREE_PATH.DESCENDANT.eq(ULong.valueOf(parentCommentNo)))
+                        .unionAll(DSL.select(DSL.val(postNo), DSL.val(ULong.valueOf(commentNo)),
                                 DSL.val(commentNo), DSL.val(0)))
                 ).execute();
     }
